+++ stablebear/__init__.py
@@ -19,9 +19,9 @@ from .base_tensor import (
     IntPcfTensor,
     IntTensor,
     PcfTensor,
-    PointCloud,
     PointCloudTensor,
 )
+from .point_cloud import PointCloud
 from .tensor_create import array_split, concatenate, split, stack, tensor, zeros
 from .typing import (
     dtype,